% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{bis_dl}
\alias{bis_dl}
\title{Download data from the PRH open data API}
\usage{
bis_dl(id, lang = "FI")
}
\arguments{
\item{id}{The business ID of the company. Must be of the format 1111111-1,
ie. seven digits followed by a dash and a checksum digit.}

\item{lang}{The language of the results. Defaults to 'FI' and must be either
'FI', 'SE' or 'EN'.}
}
\value{
A tibble with one row and the columns \code{name},
 \code{business_id}, \code{company_form}, \code{registration_date},
 \code{business_line} and \code{business_line_code}. For information about
 the business line classification, see
 \url{http://www.stat.fi/meta/luokitukset/toimiala/001-2008/index_en.html}.
}
\description{
Download information related to the given Business ID from the
Business Information System API. The language of the result can be set
with the \code{lang}-parameter.
}
\examples{
\donttest{
library(simpleprh)
# Download information related to Nokia Oyj
bis_dl("0112038-9", lang = "EN")
}

}
