% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bis_lookup.R
\name{bis_lookup}
\alias{bis_lookup}
\title{Find business information}
\usage{
bis_lookup(name, max_results = 1000L, results_from = 0L)
}
\arguments{
\item{name}{The name, or par of the name of the company.}

\item{max_results}{The maximum number of results to be returned. Note, that
the API limits this to \code{1000}.}

\item{results_from}{The starting index for the results. For example, if a
certain query returned more than \code{1000} results, having this as
\code{0} would return the first \code{1000} results and then setting this
to \code{1000} would return the next \code{1000}.}
}
\value{
A tibble with the fields \code{name}, \code{business_id},
\code{registration_date}, \code{company_form} that match the query.
}
\description{
Query company business ID, registration date, form and correct name with name
as a lookup parameter.
}
\examples{
\donttest{
library(simpleprh)
# Download information related to Nokia Oyj
bis_lookup("Nokia Oyj")
}

}
