% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleprh-package.R
\docType{package}
\name{simpleytj}
\alias{simpleytj}
\alias{simpleytj-package}
\title{SIMPLEPRH}
\description{
An R package for accessing the PRH open data API. Currently the
package supports obtaining the company name, form, business line and
registration date via the Business Information System API although
more functionality might be added in the future.
}
