% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{bis_dl}
\alias{bis_dl}
\title{Download data from the PRH open data API}
\usage{
bis_dl(id, excl_langs = character(0), return_type = "tbl")
}
\arguments{
\item{id}{The business ID of the company. Must be of the format 1111111-1,
ie. seven digits followed by a dash and a checksum digit.}

\item{excl_langs}{A character vector of the languages to be excluded from
the results. By default, some of the fields return the same information in
Finnish, Swedish and English. The value can therefore be for example
\code{c("SE", "EN")} to only return the results in Finnish or \code{"FI"}
to only return the results in English. Defaults to \code{character(0)}, ie.
the results are returned in all available languages.}

\item{return_type}{The return type. Either \code{"tbl"} for returning a one
row tibble with columns as the fields, or \code{"list"} for returning a
a list. A list is probably simpler for accessing a single result while a
tibble might be more convenient when combining multiple results. Defaults
to \code{"tbl"}.}
}
\value{
A tibble with one row where each column is a field returned by the
 API or a list with the elemnents corresponding to the fields. All
 \code{NULL}s are converted to \code{NA}s. For more information about the
 API, see \url{http://avoindata.prh.fi/ytj_en.html}. For information about
 the business line classification, see
 \url{http://www.stat.fi/meta/luokitukset/toimiala/001-2008/index_en.html}.
}
\description{
Download information related to the given Business ID from the
Business Information System API. The language of the result can be set
with the \code{lang}-parameter.
}
\examples{
\donttest{
library(simpleprh)
# Download information related to Nokia Oyj
bis_dl("0112038-9")
}

}
