% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{bis_dl}
\alias{bis_dl}
\title{Download data from the PRH open data API}
\usage{
bis_dl(id, lang = "FI")
}
\arguments{
\item{id}{The business ID of the company. Must be of the format 1111111-1,
ie. seven digits followed by a dash and a checksum digit.}

\item{lang}{The language of the results. Defaults to \code{"FI"} and must be either
\code{"FI"}, \code{"SE"} or \code{"EN"}.}
}
\value{
A tibble with one row and the columns \code{name},
 \code{business_id}, \code{company_form}, \code{registration_date},
 \code{business_line}, \code{business_line_code}, \code{liquidation_descr},
 \code{liquidation_registration} and \code{liquidation_end}.
 If the result contains multiple records of business line and/or liquidation,
 the newest (by registration date) is returned.
 \code{liquidation_descr} specifies the type of liquidation, if any has been
 active. If, in addition, \code{liquidation_end} is not \code{NA}, the
 liquidation has ended.
 For information about the business line classification, see
 \url{http://www.stat.fi/meta/luokitukset/toimiala/001-2008/index_en.html}.
}
\description{
Download information related to the given Business ID from the
Business Information System API. The language of the result can be set
with the \code{lang}-parameter.
}
\examples{
\donttest{
library(simpleprh)
# Download information related to Nokia Oyj
bis_dl("0112038-9", lang = "EN")
}

}
